Disable appending of migrations to the app.
The schema_migrations table is not updated with the migration versions
from the engine when running `rake db:setup`. This leads to a situation
where schema.rb contains the objects setup by the engine (from a
previous `rake db:migrate`), but an invocation of `rake db:migrate` on a
fresh copy of the database will fail because the engine migrations are
not recorded in schema_migrations and rails will attempt to run them
again resulting in a conflict.

This change prefers copying the migrations into the target app which
causes `rake db:setup` to add an entry to the schema_migrations
table and keeping everything in sync.